Add capability of timeframe based rotation
Introduces a new type of rotation of log files based on time frames. Any
type of timeframe expressible via time format strings can be used;
though the use of format caching prevents uses of timeframes in the
order of seconds. A known restriction is that log files are not rotated
after a restart of the application.