Parse all IPs in X-Forwarded-For header and pull out the first one for logging
The X-Forwarded-For header has a "standard" format:
https://en.wikipedia.org/wiki/X-Forwarded-For

Here's an example value you may see:

```
X-Forwarded-For: 203.0.113.195, 70.41.3.18, 150.172.238.178
```

This would be a request that has passed through 3 proxies,
where the client's IP address is 203.0.113.195.

This commit modifies the Apache logger so that only the first IP address
from this X-Forwarded-For list is written to the Apache log.